Moved validations into utils
