+++ src/Widgets/FormButtonWidget/FormButtonWidgetEditingConfig.js
@@ -1,6 +1,6 @@
 import * as Scrivito from "scrivito";
 import formButtonWidgetIcon from "../../assets/images/form_button_widget.svg";
-import { insideFormContainerValidation } from "../FormContainerWidget/validations/insideFormContainerValidation";
+import { insideFormContainerValidation } from "../FormContainerWidget/utils/validations/insideFormContainerValidation";
 
 Scrivito.provideEditingConfig("FormButtonWidget", {
   title: "Form Button",
