+++ src/Widgets/FormInputFieldWidget/FormInputFieldWidgetEditingConfig.js
@@ -1,9 +1,9 @@
 import * as Scrivito from "scrivito";
 import formInputFieldWidgetIcon from "../../assets/images/form_input_field_widget.svg";
 import { isCustomType } from "../FormContainerWidget/utils/isCustomType";
-import { customFieldNameValidation } from "../FormContainerWidget/validations/customFieldNameValidation";
-import { insideFormContainerValidation } from "../FormContainerWidget/validations/insideFormContainerValidation";
-import { typeValidation } from "../FormContainerWidget/validations/typeValidation";
+import { customFieldNameValidation } from "../FormContainerWidget/utils/validations/customFieldNameValidation";
+import { insideFormContainerValidation } from "../FormContainerWidget/utils/validations/insideFormContainerValidation";
+import { typeValidation } from "../FormContainerWidget/utils/validations/typeValidation";
 
 Scrivito.provideEditingConfig("FormInputFieldWidget", {
   title: "Form Input Field",
